# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 车载视觉感知与人机交互系统 API
  version: 1.0.0
  description: |
    覆盖车牌识别、交警手势识别、车主手势控车、日志监控与告警智能体四个模块。
    Swagger UI：`/api/docs`；原始运行时定义：`/api/openapi.json`。
servers:
  - url: http://{host}:8001
    variables:
      host:
        default: localhost
...
paths:
...
  /api/owner-gesture/recognize:
    post:
      tags: [车主手势控车]
      summary: 识别车主手势，并在存在动作时更新模拟车辆状态
      security: [{bearerAuth: []}]
      requestBody: {$ref: '#/components/requestBodies/ImageFile'}
      responses:
        '200': {description: 识别及控车完成, content: {application/json: {schema: {$ref: '#/components/schemas/GestureResponse'}}}}
        '500': {$ref: '#/components/responses/RecognitionFailure'}
  /api/owner-gesture/vehicle-state:
    get:
      tags: [车主手势控车]
      summary: 获取当前用户的模拟车辆状态
      security: [{bearerAuth: []}]
      responses:
        '200': {description: 当前车辆状态, content: {application/json: {schema: {$ref: '#/components/schemas/VehicleState'}}}}
    put:
      tags: [车主手势控车]
      summary: 手动更新当前用户的模拟车辆状态
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/VehicleState'}}}
      responses:
        '200': {description: 更新后的车辆状态, content: {application/json: {schema: {$ref: '#/components/schemas/VehicleState'}}}}
        '422': {$ref: '#/components/responses/ValidationError'}
...
  /api/owner-gesture/history:
    get:
      tags: [车主手势控车]
      summary: 查询车主手势识别历史
      parameters: [{$ref: '#/components/parameters/Skip'}, {$ref: '#/components/parameters/HistoryLimit'}]
      responses:
        '200': {description: 历史记录, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/GestureHistoryItem'}}}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    Skip: {name: skip, in: query, schema: {type: integer, minimum: 0, default: 0}}
    HistoryLimit: {name: limit, in: query, schema: {type: integer, minimum: 1, default: 20}}
  requestBodies:
    ImageFile:
      required: true
      content: {multipart/form-data: {schema: {type: object, required: [file], properties: {file: {type: string, format: binary, description: 图片或 GIF 文件}}}}}
...
  responses:
...
    ValidationError: {description: 请求校验失败, content: {application/json: {schema: {$ref: '#/components/schemas/ValidationError'}}}}
    RecognitionFailure: {description: 图像/模型处理失败, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
...
    GestureResponse: {type: object, required: [gesture, gesture_cn, confidence, annotated_image, keypoints, success], properties: {gesture: {type: string}, gesture_cn: {type: string}, confidence: {type: number, format: float}, annotated_image: {type: string, description: JPEG 的 Base64 编码}, keypoints: {type: array, items: {$ref: '#/components/schemas/PoseKeypoint'}}, success: {type: boolean}, record_id: {type: integer, nullable: true}, action: {type: string, nullable: true, description: 仅车主手势控车可能返回}}}
    PoseKeypoint: {type: object, properties: {id: {type: integer}, x: {type: number}, y: {type: number}, z: {type: number}, visibility: {type: number}}}
...
    GestureHistoryItem: {type: object, properties: {id: {type: integer}, gesture: {type: string}, gesture_cn: {type: string}, confidence: {type: number}, action: {type: string, nullable: true}, annotated_image: {type: string, nullable: true}, created_at: {type: string, format: date-time}}}
    VehicleState: {type: object, required: [volume, temperature, phone_status, current_page, is_awake], properties: {volume: {type: integer}, temperature: {type: integer}, phone_status: {type: string}, current_page: {type: string}, is_awake: {type: integer, enum: [0, 1]}}}
...
    Error: {type: object, required: [detail], properties: {detail: {type: string, description: FastAPI 当前实现的错误消息}}}
    ValidationError: {type: object, required: [detail], properties: {detail: {type: array, items: {type: object, additionalProperties: true}}}}